+++ bika/lims/browser/client/__init__.py
@@ -30,6 +30,3 @@ from views.orders import ClientOrdersView
 from views.contacts import ClientContactsView
 from views.contacts import ClientContactVocabularyFactory
 from views.samplingrounds import ClientSamplingRoundsView
-
-from ajax import ReferenceWidgetVocabulary
-from ajax import ajaxGetClientInfo