+++ bika/lims/content/arreport.py
@@ -26,6 +26,7 @@ from bika.lims.browser.widgets import DateTimeWidget
 from bika.lims.browser.widgets import ReferenceWidget
 from bika.lims.config import PROJECTNAME
 from bika.lims.content.bikaschema import BikaSchema
+from bika.lims.content.clientawaremixin import ClientAwareMixin
 from bika.lims.interfaces import IARReport
 from plone.app.blob.field import BlobField
 from Products.Archetypes import atapi
@@ -136,7 +137,7 @@ schema["id"].required = False
 schema["title"].required = False
 
 
-class ARReport(BaseFolder):
+class ARReport(BaseFolder, ClientAwareMixin):
     """An AnalysisRequest report, containing the report itself in pdf and html
        format. It includes information about the date when was published, from
        whom, the report recipients (and their emails) and the publication mode