+++ bika/lims/interfaces/__init__.py
@@ -984,3 +984,18 @@ class IAddSampleObjectInfo(Interface):
         - field_values: contains default values for other fields to be applied
           when the value of the current field changes.
         """
+
+
+class IClientAwareMixin(Interface):
+    """Marker interface for objects that can be bound to a Client, either
+    because they can be added inside a Client folder or because it can be
+    assigned through a Reference field
+    """
+
+    def getClient(self):
+        """Returns the client this object is bound to, if any
+        """
+
+    def getClientUID(self):
+        """Returns the client UID this object is bound to, if any
+        """